backend/requirements.txt
# FoodVision Backend Dependencies
# Python 3.9+

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# Database
PyMySQL==1.1.0
mysql-connector-python==8.1.0

# API Clients
google-generativeai==0.3.2
requests==2.31.0
httpx==0.25.0

# Image Processing
Pillow==10.3.0
opencv-python==4.8.1.78

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Validation
marshmallow==3.20.1
cerberus==1.3.5

# File Upload
werkzeug==2.3.7

# Date/Time
python-dateutil==2.8.2

# Development Tools
python-decouple==3.8

# Testing (optional)
pytest==7.4.2
pytest-flask==1.3.0

# Production WSGI Server
gunicorn==21.2.0

# Monitoring (optional)
sentry-sdk==1.32.0
